test: split large transaction tests
